refactor: nitpick.files to nitpick.plugins
