--- docs/source/nitpick.files.base.rst
@@ -1,7 +1,0 @@
-nitpick.files.base module
-=========================
-
-.. automodule:: nitpick.files.base
-   :members:
-   :undoc-members:
-   :show-inheritance:
