--- docs/source/nitpick.files.json.rst
@@ -1,7 +1,0 @@
-nitpick.files.json module
-=========================
-
-.. automodule:: nitpick.files.json
-   :members:
-   :undoc-members:
-   :show-inheritance:
