--- docs/source/nitpick.files.pre_commit.rst
@@ -1,7 +1,0 @@
-nitpick.files.pre\_commit module
-================================
-
-.. automodule:: nitpick.files.pre_commit
-   :members:
-   :undoc-members:
-   :show-inheritance:
