--- docs/source/nitpick.files.pyproject_toml.rst
@@ -1,7 +1,0 @@
-nitpick.files.pyproject\_toml module
-====================================
-
-.. automodule:: nitpick.files.pyproject_toml
-   :members:
-   :undoc-members:
-   :show-inheritance:
