--- docs/source/nitpick.files.rst
@@ -1,20 +1,0 @@
-nitpick.files package
-=====================
-
-.. automodule:: nitpick.files
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 4
-
-   nitpick.files.base
-   nitpick.files.json
-   nitpick.files.pre_commit
-   nitpick.files.pyproject_toml
-   nitpick.files.setup_cfg
-   nitpick.files.text
