--- docs/source/nitpick.files.setup_cfg.rst
@@ -1,7 +1,0 @@
-nitpick.files.setup\_cfg module
-===============================
-
-.. automodule:: nitpick.files.setup_cfg
-   :members:
-   :undoc-members:
-   :show-inheritance:
