--- docs/source/nitpick.plugin.rst
@@ -1,7 +1,0 @@
-nitpick.plugin module
-=====================
-
-.. automodule:: nitpick.plugin
-   :members:
-   :undoc-members:
-   :show-inheritance:
