+++ docs/source/nitpick.plugins.base.rst
@@ -1,0 +1,7 @@
+nitpick.plugins.base module
+===========================
+
+.. automodule:: nitpick.plugins.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
