+++ docs/source/nitpick.plugins.json.rst
@@ -1,0 +1,7 @@
+nitpick.plugins.json module
+===========================
+
+.. automodule:: nitpick.plugins.json
+   :members:
+   :undoc-members:
+   :show-inheritance:
