+++ docs/source/nitpick.plugins.pre_commit.rst
@@ -1,0 +1,7 @@
+nitpick.plugins.pre\_commit module
+==================================
+
+.. automodule:: nitpick.plugins.pre_commit
+   :members:
+   :undoc-members:
+   :show-inheritance:
