+++ docs/source/nitpick.plugins.rst
@@ -1,0 +1,19 @@
+nitpick.plugins package
+=======================
+
+.. automodule:: nitpick.plugins
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   nitpick.plugins.base
+   nitpick.plugins.json
+   nitpick.plugins.pre_commit
+   nitpick.plugins.pyproject_toml
+   nitpick.plugins.setup_cfg
