+++ docs/source/nitpick.plugins.setup_cfg.rst
@@ -1,0 +1,7 @@
+nitpick.plugins.setup\_cfg module
+=================================
+
+.. automodule:: nitpick.plugins.setup_cfg
+   :members:
+   :undoc-members:
+   :show-inheritance:
