+++ docs/source/nitpick.rst
@@ -12,7 +12,7 @@ Subpackages
 .. toctree::
    :maxdepth: 4
 
-   nitpick.files
+   nitpick.plugins
 
 Submodules
 ----------
@@ -29,7 +29,6 @@ Submodules
    nitpick.formats
    nitpick.generic
    nitpick.mixin
-   nitpick.plugin
    nitpick.schemas
    nitpick.style
    nitpick.typedefs
