+++ pyproject.toml
@@ -28,10 +28,10 @@ classifiers = [
 NIP = "nitpick.flake8:NitpickChecker"
 
 [tool.poetry.plugins.nitpick]
-json = "nitpick.files.json"
-pre_commit = "nitpick.files.pre_commit"
-setup_cfg = "nitpick.files.setup_cfg"
-pyproject_toml = "nitpick.files.pyproject_toml"
+json = "nitpick.plugins.json"
+pre_commit = "nitpick.plugins.pre_commit"
+setup_cfg = "nitpick.plugins.setup_cfg"
+pyproject_toml = "nitpick.plugins.pyproject_toml"
 
 [tool.poetry.dependencies]
 python = "^3.5 || ^3.6 || ^3.7 || ^3.8"
