+++ src/nitpick/constants.py
@@ -30,5 +30,3 @@ TOOL_NITPICK = "tool.nitpick"
 TOOL_NITPICK_JMEX = jmespath.compile(TOOL_NITPICK)
 NITPICK_STYLES_INCLUDE_JMEX = jmespath.compile("nitpick.styles.include")
 NITPICK_MINIMUM_VERSION_JMEX = jmespath.compile("nitpick.minimum_version")
-
-KEY_FILE_NAMES = "file_names"
