--- src/nitpick/files/__init__.py
@@ -1,1 +1,0 @@
-"""Files that are checked by Nitpick."""
