+++ src/nitpick/plugins/__init__.py
@@ -1,12 +1,13 @@
-"""Flake8 plugin to check files."""
-from typing import TYPE_CHECKING, Any, Dict, Optional, Set
+"""Hooks used by Nitpick pkugins."""
+from typing import TYPE_CHECKING, Optional, Set
 
 import pluggy
 
 from nitpick.constants import PROJECT_NAME
+from nitpick.typedefs import JsonDict
 
 if TYPE_CHECKING:
-    from nitpick.files.base import BaseFile
+    from nitpick.plugins.base import BaseFile
 
 
 hookspec = pluggy.HookspecMarker(PROJECT_NAME)
@@ -15,6 +16,6 @@ hookimpl = pluggy.HookimplMarker(PROJECT_NAME)
 
 @hookspec
 def handle_config_file(  # pylint: disable=unused-argument
-    filename: str, tags: Set[str], config_dict: Dict[str, Any]
+    config: JsonDict, file_name: str, tags: Set[str]
 ) -> Optional["BaseFile"]:
     """Return a BaseFile if this plugin handles the relative filename or any of its :py:package:`identify` tags."""
