+++ src/nitpick/plugins/pyproject_toml.py
@@ -1,10 +1,10 @@
 """Checker for `pyproject.toml <https://github.com/python-poetry/poetry/blob/master/docs/docs/pyproject.md>`_."""
-from typing import Any, Dict, Optional, Set
+from typing import Optional, Set
 
 from nitpick.app import Nitpick
-from nitpick.files.base import BaseFile
-from nitpick.plugin import hookimpl
-from nitpick.typedefs import YieldFlake8Error
+from nitpick.plugins import hookimpl
+from nitpick.plugins.base import BaseFile
+from nitpick.typedefs import JsonDict, YieldFlake8Error
 
 
 class PyProjectTomlFile(BaseFile):
@@ -32,8 +32,8 @@ class PyProjectTomlFile(BaseFile):
 
 @hookimpl
 def handle_config_file(  # pylint: disable=unused-argument
-    filename: str, tags: Set[str], config_dict: Dict[str, Any]
+    config: JsonDict, file_name: str, tags: Set[str]
 ) -> Optional["BaseFile"]:
     """Handle pyproject.toml file."""
-    base_file = PyProjectTomlFile()
-    return base_file if filename == base_file.file_name else None
+    base_file = PyProjectTomlFile(config)
+    return base_file if file_name == base_file.file_name else None
