+++ tests/helpers.py
@@ -10,11 +10,11 @@ from testfixtures import compare
 
 from nitpick.app import Nitpick
 from nitpick.constants import CACHE_DIR_NAME, ERROR_PREFIX, MERGED_STYLE_TOML, NITPICK_STYLE_TOML, PROJECT_NAME
-from nitpick.files.pre_commit import PreCommitFile
-from nitpick.files.pyproject_toml import PyProjectTomlFile
-from nitpick.files.setup_cfg import SetupCfgFile
 from nitpick.flake8 import NitpickChecker
 from nitpick.formats import TomlFormat
+from nitpick.plugins.pre_commit import PreCommitFile
+from nitpick.plugins.pyproject_toml import PyProjectTomlFile
+from nitpick.plugins.setup_cfg import SetupCfgFile
 from nitpick.typedefs import PathOrStr
 from tests.conftest import TEMP_ROOT_PATH
